Only logged in users can add/delete pins, added relevant error pages.

+++ routes/maps.js
@@ -24,7 +24,7 @@ module.exports = (db) => {
     const mapID = req.params.id;
     const templateVars = {};
     if (!mapID) {
-      res.statusCode = 404; //create views/404.ejs
+      res.statusCode = 404;
       if (!req.session.user_id) {
         templateVars.user = null;
         templateVars.id = null;
@@ -52,32 +52,44 @@ module.exports = (db) => {
 
 // Add a new pin to db
   router.post("/", (req, res) => {
-    let queryString = `
-      INSERT INTO pins (title, description, image_url, latitude, longitude, created_at, map_id)
-      VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *;
-    `;
-    // req.body = {} from the submit form for new pin
-    let values = req.body;
-    let { title, description, image_url, latitude, longitude, map_id } = values;
+    const templateVars = {};
+    if (!req.session.user_id) {
+      templateVars.user = null;
+      templateVars.id = null;
+      res.statusCode = 401;
+      res.render('401', templateVars);
+    } else {
+      templateVars.user = req.session.username;
+      templateVars.id = req.session.user_id;
+      let queryString = `
+        INSERT INTO pins (title, description, image_url, latitude, longitude, created_at, map_id, user_id)
+        VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING *;
+      `;
+      // req.body = {} from the submit form for new pin
+      let user_id = req.session.user_id;
+      let values = req.body;
+      let { title, description, image_url, latitude, longitude, map_id } = values;
 
-    console.log("values from .post", values);
+      console.log("values from .post", values);
 
-    db.query(queryString, [
-      title,
-      description,
-      image_url,
-      parseFloat(latitude),
-      parseFloat(longitude),
-      new Date(),
-      map_id
-    ])
-      .then((data) => {
-        console.log("data.rows from post", data.rows);
-        return data.rows;
-      })
-      .catch((err) => console.error("query error", err.stack));
-    // console.log([values.title, values.description]);
-    res.redirect("/maps");
+      db.query(queryString, [
+        title,
+        description,
+        image_url,
+        parseFloat(latitude),
+        parseFloat(longitude),
+        new Date(),
+        map_id,
+        user_id
+      ])
+        .then((data) => {
+          console.log("data.rows from post", data.rows);
+          return data.rows;
+        })
+        .catch((err) => console.error("query error", err.stack));
+      // console.log([values.title, values.description]);
+      res.redirect("/maps");
+    }
   });
 
   // instead of res.redirect which refreshes the page, do res.json(data.rows[0]) to the popup with AJAX
@@ -86,16 +98,26 @@ module.exports = (db) => {
    // Remove pins from maps.js
   router.post("/:id/delete", (req, res) => {
     console.log("id from maps.js delete:", req.params.id);
-    let values = [req.params.id];
-    let queryString = `
-      DELETE FROM pins
-      WHERE id = $1;`;
+    const templateVars = {};
+    if (!req.session.user_id) {
+      templateVars.user = null;
+      templateVars.id = null;
+      res.statusCode = 401;
+      res.render('401', templateVars);
+    } else {
+      templateVars.user = req.session.username;
+      templateVars.id = req.session.user_id;
+      let values = [req.params.id];
+      let queryString = `
+        DELETE FROM pins
+        WHERE id = $1;`;
 
-    db.query(queryString, values)
-      .then((data) => console.log(data))
-      .catch((err) => console.lerror("query error", err.stack));
+      db.query(queryString, values)
+        .then((data) => console.log(data))
+        .catch((err) => console.lerror("query error", err.stack));
 
-    res.redirect("/maps");
+      res.redirect("/maps");
+    }
   });
 
   return router;
